requirements.txt
amqp==5.1.1
apturl==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
boto3==1.24.80
botocore==1.27.80
Brlapi==0.8.3
CacheControl==0.12.11
cachetools==5.2.0
carehare==1.0.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-rabbitmq==4.0.1
channels-redis==4.0.0
chardet==4.0.0
charset-normalizer==2.1.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.1
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
distro-info===1.1build1
dj-database-url==1.0.0
Django==4.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-oauth-toolkit==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-extra-fields==3.4.0
drf-writable-nested==0.7.0
drf-yasg==1.21.3
duplicity==0.8.21
eventlet==0.33.1
exceptiongroup==1.0.4
execnet==1.9.0
fasteners==0.14.1
fcm-django==1.0.12
firebase-admin==5.4.0
future==0.18.2
gevent==22.10.2
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio==1.49.1
grpcio-status==1.49.1
gssapi==1.8.2
gunicorn==20.1.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.9.1
jwcrypto==1.4.2
keyring==23.5.0
kombu==5.2.4
krb5==0.4.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
model-bakery==1.9.0
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.4
netifaces==0.11.0
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pamqp==3.2.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.32
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.9.4
psycopg2==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.4.0
pymacaroons==0.13.0
pymongo==4.2.0
pymongo-auth-aws==1.1.0
pymongocrypt==1.3.1
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
pyspnego==0.6.0
pytest==7.2.0
pytest-django==4.5.2
pytest-parallel==0.1.1
pytest-xdist==3.1.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-snappy==0.6.1
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
redis==4.3.4
reportlab==3.6.8
requests==2.28.1
requests-kerberos==0.14.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
SecretStorage==3.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
supervisor==4.2.1
swagger-spec-validator==2.7.4
systemd-python==234
tblib==1.7.0
texttable==1.6.4
tomli==2.0.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.1
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.11
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.2.3
whitenoise==6.2.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.5.0
zstandard==0.18.0
